LLM/requirements.txt
transformers==4.43.0
datasets
accelerate
hydra-core
jsonlines
nltk
numpy>=1.23,<2
omegaconf
peft
PyYAML
rouge_score
scikit_learn
scipy
sentence_transformers>=2.7,<3.0
tqdm

